fixed wrong implementation of delete method
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@3344 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ content/classes/class.ilObjContentObject.php
@@ -435,20 +435,25 @@ class ilObjContentObject extends ilObject
 	*/
 	function delete()
 	{
+		// always call parent delete function first!!
+		if (!parent::delete())
+		{
+			return false;
+		}
+		
 		// delete lm object data
-		require_once("content/classes/class.ilLMObject.php");
+		include_once("content/classes/class.ilLMObject.php");
 		ilLMObject::_deleteAllObjectData($this);
 
 		// delete meta data of content object
 		$nested = new ilNestedSetXML();
 		$nested->init($this->getId(), $this->getType());
 		$nested->deleteAllDBData();
-//echo "2";
+
 		// delete learning module tree
 		$this->lm_tree->removeTree($this->lm_tree->getTreeId());
-//echo "3";
-		// always call parent delete function at the end!!
-		return (parent::delete()) ? true : false;
+
+		return true;
 	}
 
 	/**
